merge develop a main

+++ database/migrations/2025_07_14_112018_add_evento_id_to_seguimientos_table.php
@@ -8,13 +8,14 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::table('seguimientos', function (Blueprint $table) {
-            // agregar columna
-            $table->foreignId('evento_id')
-                ->nullable()
-                ->constrained('eventos')
-                ->onDelete('set null');
-        });
+        if (! Schema::hasColumn('seguimientos', 'evento_id')) {
+            Schema::table('seguimientos', function (Blueprint $table) {
+                $table->foreignId('evento_id')
+                    ->nullable()
+                    ->constrained('eventos')
+                    ->onDelete('set null');
+            });
+        }
     }
 
     public function down()
